fmt/zone: skip OPT record in sizes calculation
OPT record's fields are not as of a regular record.
And OPT record is not formatted as other records in zone file format.
Thus it must be skipped in sizes calculation. Otherwise field width is
calculated based on a non-printed field (as OPT is skipped).

In particular, this affects field size calculations because OPT record's
CLASS denotes the UDP payload size. Thus, this value is printed as
CLASS65535 (in case of 64K payload). This generates a field size of 11
characters which is excessive when all printed records have CLASS = IN
which needs the regular field width.